+++ spec/provider/package/apt_spec.rb
@@ -7,13 +7,17 @@ describe Wright::Provider::Package::Apt do
     ['apt-cache', cmd.to_s, pkg_name]
   end
 
-  def apt_get(action, pkg_name, pkg_version = nil)
-    version = pkg_version.nil? ? '' : "=#{pkg_version}"
-    ['apt-get', action.to_s, '-qy', pkg_name + version]
+  def apt_get(action, pkg_name, args = {})
+    version = args[:version].nil? ? '' : "=#{args[:version]}"
+    options = args[:options]
+    ['apt-get', *options, action.to_s, '-qy', pkg_name + version]
   end
 
-  def package_provider(pkg_name, pkg_version = nil)
-    pkg_resource = OpenStruct.new(name: pkg_name, version: pkg_version)
+  def package_provider(pkg_name, args = {})
+    properties = { name: pkg_name,
+                   version: args[:version],
+                   options: args[:options]}
+    pkg_resource = OpenStruct.new(properties)
     Wright::Provider::Package::Apt.new(pkg_resource)
   end
 
@@ -116,10 +120,22 @@ describe Wright::Provider::Package::Apt do
     it 'should install specific package versions' do
       pkg_name = 'abcde'
       pkg_version = '2.5.4-1'
-      pkg_provider = package_provider(pkg_name, pkg_version)
-      apt_cmd = apt_get(:install, pkg_name, pkg_version)
+      pkg_provider = package_provider(pkg_name, version: pkg_version)
+      apt_cmd = apt_get(:install, pkg_name, version: pkg_version)
 
       @fake_capture3.expect(apt_cmd)
+      @fake_capture3.stub do
+        pkg_provider.send(:install_package)
+      end
+    end
+
+    it 'should pass options to the package manager' do
+      pkg_name = 'htop'
+      pkg_options = ['--no-install-recommends']
+      pkg_provider = package_provider(pkg_name, options: pkg_options)
+      apt_cmd = apt_get(:install, pkg_name, options: pkg_options)
+
+      @fake_capture3.expect(apt_cmd, 'apt-get_install_-qy_htop')
       @fake_capture3.stub do
         pkg_provider.send(:install_package)
       end
@@ -151,5 +167,18 @@ describe Wright::Provider::Package::Apt do
         pkg_provider.send(:remove_package)
       end
     end
+
+    it 'should pass options to the package manager' do
+      pkg_name = 'abcde'
+      pkg_options = ['--purge']
+      pkg_provider = package_provider(pkg_name, options: pkg_options)
+      apt_cmd = apt_get(:remove, pkg_name, options: pkg_options)
+
+      @fake_capture3.expect(apt_cmd)
+      @fake_capture3.stub do
+        pkg_provider.send(:remove_package)
+      end
+    end
+
   end
 end